feat(engine-biorhythm): extract calculation functions into calculator.rs module (issue #369)
- Created calculator.rs with all cycle constants, result types (BiorhythmResult,
  CycleResult, ForecastDay), and every pure math function (cycle_value,
  to_percentage, phase_label, days_until_peak, days_until_critical,
  is_critical_day, compute_cycle, find_critical_days, build_forecast,
  generate_witness_prompt)
- Internal trigonometric helpers are pub(crate); higher-level API functions
  and types remain pub
- Refactored lib.rs to declare pub mod calculator, re-export only public items,
  and contain only BiorhythmEngine + ConsciousnessEngine impl + parse_date
- All 12 existing tests pass unchanged
- Deleted .agent-tasks/issue-369.md

Agent-Logs-Url: https://github.com/Sheshiyer/Selemene-engine/sessions/ce0e77fa-527a-4d33-aead-8d3e9ae1a950
