--- .agent-tasks/issue-369.md
@@ -1,8 +1,0 @@
-# Agent Task: Issue #369
-
-Title: [V22-W1-S2-02] Extract calculation functions into calculator.rs module
-Issue: https://github.com/Sheshiyer/Selemene-engine/issues/369
-
-Implement the issue deliverable and acceptance criteria.
-Run cargo build and cargo test before marking ready.
-Delete this file before merge.
